Logo works, still trying to get a Searchbar to nav. HTML-File for the history is called. and called correctly. but still empty and i dont have aclue right now //This is fine now. Now I have to find out how to edit the Text in the database via a WYSIWYG Editor for the User AAAAAAAND I can finally display Items on the right side of the navigationbar. I can register and delete and Save users with hashed passwords! Trying to implement or get this wysiwyg Editor to do what I want!! first steps toward an Upload for Images //seems not to make any sense because if the wysiwyg works this will be no longer needed except for maybe creating a header picture for every single board. Login Logout works with registered users. but I dont know right now how secure this is :) I hope, I stilll know what I did //Nope by now, next Time I will continue my thougts
Okay, I can add History to Boards and I can add boards and read out the History and display it very beatuiful. It is kinda cool how the new History objects are added. so I still need to implement the editing of the History and everything else..
so far now.

good night!

Editing is not working yet- yes it is :)

Searching from NAv an see TODO