Need User Profile page for changing passwords and stuff
Tried to change color of targeted div, only working after refresh because the css is not reloaded i guess

If more problems, change primary key back to username and User.get_id back so that it returns the username as well as User.get()

So far everything works smooth af. I need to get this thing testet so I find some bugs.